requirements.txt
annotated-types
anyio
bcrypt
click
colorama
ecdsa
fastapi
h11
idna
mysql-connector-python
names
numpy
pandas
passlib
pyasn1
pydantic
pydantic_core
PyMySQL
pyotp
python-dateutil
python-dotenv
python-jose
python-multipart
pytz
random-password-generator
rsa
six
sniffio
starlette
typing_extensions
tzdata
uvicorn
aiosmtplib
pydantic-settings
aiomysql
pytest
pytest-asyncio
boto3